requirements.txt
-e .
jupyter
matplotlib
numpy
pymongo==3.10.1
sacred>=0.8.1
scipy
tensorflow-datasets>=1.3.2
tensorflow-gpu==2.1.0
tensorflow-probability==0.9.0
tqdm
cython
